requirements.txt
python-jose
fastapi
bcrypt
pydantic
pydantic-settings
uvicorn
SQLAlchemy
python-multipart
